requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
dataclasses-json==0.6.7
distro==1.9.0
docopt==0.6.2
Flask==2.3.3
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-deepseek==0.1.3
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
loguru==0.7.3
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
pipwin==0.5.2
propcache==0.3.1
PyAudio==0.2.14
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyjsparser==2.7.1
PyJWT==2.10.1
PyPDF2==3.0.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tk==0.1.0
tqdm==4.67.1
ttkthemes==3.2.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.0.1
win32_setctime==1.2.0
yarl==1.20.0
zstandard==0.23.0
